+++ cdk/lib/classical-music-lake-stack.ts
@@ -420,7 +420,7 @@ export class ClassicalMusicLakeStack extends cdk.Stack {
     const authVerifyEmail = fn("AuthVerifyEmail", "handlers/auth/verify-email.ts");
     const authResendCode = fn("AuthResendCode", "handlers/auth/resend-verification-code.ts");
     const authRefresh = fn("AuthRefresh", "handlers/auth/refresh.ts");
-    const authPreSignUp = fn("AuthPreSignUp", "handlers/auth/pre-signup.ts");
+    const authPreSignUp = fn("AuthPreSignUp", "handlers/cognito/pre-signup.ts");
 
     const concertLogsList = fn("ConcertLogsList", "handlers/concert-logs/list.ts");
     const concertLogsCreate = fn("ConcertLogsCreate", "handlers/concert-logs/create.ts");
